+++ CHANGELOG.md
@@ -1,7 +1,14 @@
 # Latest changes
-## v2.6.13
-### Bugfixes
-- Fixed a bug where collage cover headers would erroneously have their colors changed to the NSE theme color
+## v2.7
+### Features
+On Torrent Detail pages:
+- Right-Click Management now works for both uploader elements, not just the top one
+- The seeders count will now be highlighted in red if you have NSE set to hide unseeded torrents and there are no seeders
+- Both uploader and title fields will now highlight in the same manner as browse pages when a filtered uploader or title phrase was detected
+
+Global:
+- Most elements that are modified due to the torrent being (un)filtered now have a tooltip telling you why it was (un)filtered
+
 
 ---